# app/backend/api/openapi.yml
openapi: 3.1.0
info:
  title: Orz People Platform API
  description: HTTP API for the Orz People Platform.
  version: 0.1.0
servers:
  - url: http://localhost:6770
    description: Local development
...
paths:
...
  /forms/getPublishedForm/{formId}:
    get:
      tags: [Form filling]
      summary: Get the active published Form definition by global Form ID
      operationId: getPublishedForm
      description: >-
        Authentication is optional. If a Bearer token is supplied it must be valid;
        authenticated update-subject callers may receive their actor-bound context.
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      responses:
        '200':
          description: Minimal active published definition and current availability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishedFormEnvelope'
        '401':
          description: A supplied Bearer token is invalid or no longer has a live Session.
        '404':
          description: The Form is missing, archived, or has no active published release.
  /forms/getRelationOptions/{formId}/{itemId}:
    get:
      tags: [Form filling]
      summary: Get safe dynamic options for a published relation item
      operationId: getFormRelationOptions
      security: []
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: values
          in: query
          description: JSON object containing current declared valueFrom dependencies.
          schema:
            type: string
            maxLength: 20000
        - name: take
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
      responses:
        '200':
          description: Opaque row IDs and display labels only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormRelationOptionsEnvelope'
        '400':
          description: Invalid values/take or the item is not a configured relation selector.
        '404':
          description: The published Form or item was not found.
  /forms/submitForm:
    post:
      tags: [Form filling]
      summary: Submit answers to a published Form
      operationId: submitForm
      description: >-
        Authentication is optional. The operation revalidates availability, access,
        Schema, Dataset mappings, relations and subject-row revision on the server.
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitFormRequest'
      responses:
        '201':
          description: Submission created or subject row updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitFormEnvelope'
        '400':
          description: Invalid DTO, Schema answers, hidden item, mapping, or idempotency key.
        '401':
          description: Authentication is required or a supplied Bearer token is invalid.
        '404':
          description: The published Form or required subject row was not found.
        '409':
          description: Closed state, stale revision, or conflicting idempotency payload.
        '429':
          description: The per-Form actor/network submission limit was exceeded.
  /workspaces/{workspaceId}/datasets:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [Datasets]
      operationId: listDatasets
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Visible Datasets and creation capability.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetListEnvelope' }
    post:
      tags: [Datasets]
      operationId: createDataset
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDatasetRequest' }
      responses:
        '201':
          description: Dataset created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetEnvelope' }
  /workspaces/{workspaceId}/datasets/{datasetId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DatasetId'
    get:
      tags: [Datasets]
      operationId: getDataset
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Dataset detail, active fields, creator and capabilities.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetDetailEnvelope' }
    patch:
      tags: [Datasets]
      operationId: updateDataset
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDatasetRequest' }
      responses:
        '200':
          description: Dataset updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetEnvelope' }
  /workspaces/{workspaceId}/datasets/{datasetId}/archive:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DatasetId'
    post:
      tags: [Datasets]
      operationId: archiveDataset
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RevisionRequest' }
      responses:
        '201':
          description: Dataset archived.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetEnvelope' }
  /workspaces/{workspaceId}/datasets/{datasetId}/fields:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DatasetId'
    post:
      tags: [Datasets]
      operationId: createDatasetField
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDatasetFieldRequest' }
      responses:
        '201':
          description: Field created and positioned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetFieldMutationEnvelope' }
  /workspaces/{workspaceId}/datasets/{datasetId}/fields/{fieldId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DatasetId'
      - $ref: '#/components/parameters/FieldId'
    patch:
      tags: [Datasets]
      operationId: updateDatasetField
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDatasetFieldRequest' }
      responses:
        '200':
          description: Field updated and positions normalized.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetFieldMutationEnvelope' }
  /workspaces/{workspaceId}/datasets/{datasetId}/fields/{fieldId}/archive:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DatasetId'
      - $ref: '#/components/parameters/FieldId'
    post:
      tags: [Datasets]
      operationId: archiveDatasetField
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArchiveDatasetFieldRequest' }
      responses:
        '201':
          description: Field archived and positions normalized.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetFieldMutationEnvelope' }
  /workspaces/{workspaceId}/datasets/{datasetId}/fields/{fieldId}/options:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DatasetId'
      - $ref: '#/components/parameters/FieldId'
      - name: search
        in: query
        schema: { type: string, maxLength: 128 }
      - name: cursor
        in: query
        schema: { type: string }
      - name: limit
        in: query
        schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      - name: selectedIds
        in: query
        schema: { type: array, maxItems: 100, items: { type: string } }
    get:
      tags: [Datasets]
      operationId: listDatasetRelationOptions
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Paged relation options with selected values retained.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetRelationOptionEnvelope' }
  /workspaces/{workspaceId}/datasets/{datasetId}/rows/query:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DatasetId'
    post:
      tags: [Datasets]
      operationId: queryDatasetRows
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DatasetWindowQueryRequest' }
      responses:
        '201':
          description: Absolute Dataset row window.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetWindowEnvelope' }
  /workspaces/{workspaceId}/datasets/{datasetId}/rows:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DatasetId'
    post:
      tags: [Datasets]
      operationId: createDatasetRow
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDatasetRowRequest' }
      responses:
        '201':
          description: Materialized Dataset row.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetRowEnvelope' }
  /workspaces/{workspaceId}/datasets/{datasetId}/rows/{rowId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DatasetId'
      - name: rowId
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Datasets]
      operationId: updateDatasetRow
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDatasetRowRequest' }
      responses:
        '200':
          description: Materialized updated Dataset row.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetRowEnvelope' }
    delete:
      tags: [Datasets]
      operationId: deleteDatasetRow
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RevisionRequest' }
      responses:
        '200':
          $ref: '#/components/responses/Accepted'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Accepted:
      description: The request was accepted without revealing account state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AcceptedEnvelope'
...
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
    DatasetId:
      name: datasetId
      in: path
      required: true
      schema: { type: string }
    FieldId:
      name: fieldId
      in: path
      required: true
      schema: { type: string }
  schemas:
    EnvelopeBase:
      type: object
      required: [status, data, timestamp]
      properties:
        status:
          type: string
          enum:
            - success
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - unknown
            - account_not_found
            - invalid_credentials
            - registration_expired
            - registration_unverified
            - username_unavailable
        timestamp:
          type: string
          format: date-time
...
    AcceptedEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            status:
              const: success
            data:
              type: object
              required: [accepted]
              properties:
                accepted:
                  const: true
...
    PublishedForm:
      type: object
      required:
        - id
        - version
        - defaultLocale
        - nameI18n
        - descriptionI18n
        - closingMessageI18n
        - opensAt
        - closesAt
        - submissionAccess
        - writeMode
        - schema
        - acceptingSubmissions
        - unavailableReason
        - submissionContext
      additionalProperties: false
      properties:
        id:
          type: string
        version:
          type: integer
          minimum: 1
        defaultLocale:
          type: string
        nameI18n:
          $ref: '#/components/schemas/LocalizedText'
        descriptionI18n:
          oneOf:
            - $ref: '#/components/schemas/LocalizedText'
            - type: 'null'
        closingMessageI18n:
          oneOf:
            - $ref: '#/components/schemas/LocalizedText'
            - type: 'null'
        opensAt:
          type: [string, 'null']
          format: date-time
        closesAt:
          type: [string, 'null']
          format: date-time
        submissionAccess:
          type: string
          enum: [anonymous_allowed, authentication_required]
        writeMode:
          type: string
          enum: [create_row, update_subject_row]
        schema:
          oneOf:
            - type: object
              additionalProperties: true
            - type: boolean
        acceptingSubmissions:
          type: boolean
        unavailableReason:
          type: [string, 'null']
          enum: [not_started, closed, inactive, subject_row_missing, null]
        submissionContext:
          oneOf:
            - $ref: '#/components/schemas/FormSubmissionContext'
            - type: 'null'
    PublishedFormEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PublishedForm'
    FormRelationOption:
      type: object
      required: [id, label]
      additionalProperties: false
      properties:
        id:
          type: string
        label:
          type: string
    FormRelationOptionsEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/FormRelationOption'
    SubmitFormRequest:
      type: object
      required: [formId, answers]
      additionalProperties: false
      properties:
        formId:
          type: string
          minLength: 1
          maxLength: 128
        answers:
          type: object
          additionalProperties: true
        expectedRevision:
          type: integer
          minimum: 1
    SubmitFormResult:
      type: object
      required: [submissionId, operation, submittedAt]
      additionalProperties: false
      properties:
        submissionId:
          type: string
        operation:
          type: string
          enum: [created, updated]
        submittedAt:
          type: string
          format: date-time
    SubmitFormEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SubmitFormResult'
...
    DatasetSummary:
      type: object
      additionalProperties: false
      required: [id, workspaceId, name, slug, description, type, status, subjectMode, revision, createdAt, updatedAt]
      properties:
        id: { type: string }
        workspaceId: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: [string, 'null'] }
        type: { type: string, enum: [standard, members, join_requests, activity_registrations] }
        status: { type: string, enum: [active, archived] }
        subjectMode: { type: string, enum: [none, single_per_user] }
        revision: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    DatasetRow:
      type: object
      additionalProperties: false
      required: [id, datasetId, values, relations, revision, createdAt, updatedAt, deletedAt]
      properties:
        id: { type: string }
        datasetId: { type: string }
        values: { type: object, additionalProperties: true }
        relations:
          type: object
          additionalProperties:
            oneOf:
              - { type: string }
              - { type: array, items: { type: string } }
        revision: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: [string, 'null'], format: date-time }
    DatasetDetail:
      type: object
      required: [dataset, fields, creator, capabilities]
      properties:
        dataset: { $ref: '#/components/schemas/DatasetSummary' }
        fields: { type: array, items: { $ref: '#/components/schemas/DatasetField' } }
        creator: { $ref: '#/components/schemas/DatasetCreator' }
        capabilities: { $ref: '#/components/schemas/DatasetCapabilities' }
    DatasetList:
      type: object
      required: [items, canCreate]
      properties:
        canCreate: { type: boolean }
        items:
          type: array
          items:
            allOf:
              - { $ref: '#/components/schemas/DatasetSummary' }
              - type: object
                required: [creator, capabilities]
                properties:
                  creator: { $ref: '#/components/schemas/DatasetCreator' }
                  capabilities: { $ref: '#/components/schemas/DatasetCapabilities' }
    CreateDatasetRequest:
      type: object
      additionalProperties: false
      required: [name, slug, type]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 64 }
        description: { type: string, maxLength: 2000 }
        type: { type: string, enum: [standard, join_requests] }
        subjectMode: { type: string, enum: [none, single_per_user], default: none }
    RevisionRequest:
      type: object
      additionalProperties: false
      required: [expectedRevision]
      properties:
        expectedRevision: { type: integer, minimum: 1 }
    UpdateDatasetRequest:
      allOf:
        - { $ref: '#/components/schemas/RevisionRequest' }
        - type: object
          properties:
            name: { type: string, minLength: 1, maxLength: 128 }
            slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 64 }
            description: { type: [string, 'null'], maxLength: 2000 }
    CreateDatasetFieldRequest:
      type: object
      additionalProperties: false
      required: [expectedDatasetRevision, key, name, kind, valueSchema]
      properties:
        expectedDatasetRevision: { type: integer, minimum: 1 }
        key: { type: string, pattern: '^[a-z][a-z0-9_]*$', maxLength: 64 }
        name: { type: string, minLength: 1, maxLength: 128 }
        description: { type: [string, 'null'] }
        kind: { $ref: '#/components/schemas/DatasetField/properties/kind' }
        valueSchema: {}
        config: { type: object, additionalProperties: true }
        required: { type: boolean }
        relationTargetDatasetId: { type: string }
        relationCardinality: { type: string, enum: [one, many] }
        position: { type: integer, minimum: 0 }
    UpdateDatasetFieldRequest:
      type: object
      additionalProperties: false
      required: [expectedDatasetRevision, expectedFieldRevision]
      properties:
        expectedDatasetRevision: { type: integer, minimum: 1 }
        expectedFieldRevision: { type: integer, minimum: 1 }
        name: { type: string, minLength: 1, maxLength: 128 }
        description: { type: [string, 'null'] }
        valueSchema: {}
        config: { type: object, additionalProperties: true }
        required: { type: boolean }
        position: { type: integer, minimum: 0 }
    ArchiveDatasetFieldRequest:
      type: object
      additionalProperties: false
      required: [expectedDatasetRevision, expectedFieldRevision]
      properties:
        expectedDatasetRevision: { type: integer, minimum: 1 }
        expectedFieldRevision: { type: integer, minimum: 1 }
...
    DatasetWindowQueryRequest:
      type: object
      additionalProperties: false
      required: [query, window]
      properties:
        query: { $ref: '#/components/schemas/DatasetTableQuery' }
        window:
          type: object
          additionalProperties: false
          required: [offset, limit]
          properties:
            offset: { type: integer, minimum: 0 }
            limit: { type: integer, minimum: 1, maximum: 100 }
        includeGroupDirectory: { type: boolean }
    CreateDatasetRowRequest:
      type: object
      additionalProperties: false
      properties:
        values: { type: object, additionalProperties: true }
        relations: { type: object, additionalProperties: true }
    UpdateDatasetRowRequest:
      allOf:
        - { $ref: '#/components/schemas/RevisionRequest' }
        - { $ref: '#/components/schemas/CreateDatasetRowRequest' }
    DatasetEnvelope:
      allOf:
        - { $ref: '#/components/schemas/EnvelopeBase' }
        - type: object
          properties: { data: { $ref: '#/components/schemas/DatasetSummary' } }
    DatasetListEnvelope:
      allOf:
        - { $ref: '#/components/schemas/EnvelopeBase' }
        - type: object
          properties: { data: { $ref: '#/components/schemas/DatasetList' } }
    DatasetDetailEnvelope:
      allOf:
        - { $ref: '#/components/schemas/EnvelopeBase' }
        - type: object
          properties: { data: { $ref: '#/components/schemas/DatasetDetail' } }
    DatasetFieldMutationEnvelope:
      allOf:
        - { $ref: '#/components/schemas/EnvelopeBase' }
        - type: object
          properties:
            data:
              type: object
              required: [field, datasetRevision]
              properties:
                field: { $ref: '#/components/schemas/DatasetField' }
                datasetRevision: { type: integer, minimum: 1 }
    DatasetRowEnvelope:
      allOf:
        - { $ref: '#/components/schemas/EnvelopeBase' }
        - type: object
          properties: { data: { $ref: '#/components/schemas/DatasetRow' } }
    DatasetWindowEnvelope:
      allOf:
        - { $ref: '#/components/schemas/EnvelopeBase' }
        - type: object
          properties:
            data:
              type: object
              required: [queryFingerprint, totalRowCount, startIndex, items]
              properties:
                queryFingerprint: { type: string }
                totalRowCount: { type: integer, minimum: 0 }
                startIndex: { type: integer, minimum: 0 }
                items: { type: array, items: { $ref: '#/components/schemas/DatasetRow' } }
                groups: { type: array, items: { type: object, additionalProperties: true } }
    DatasetRelationOptionEnvelope:
      allOf:
        - { $ref: '#/components/schemas/EnvelopeBase' }
        - type: object
          properties:
            data:
              type: object
              required: [items, nextCursor]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    additionalProperties: false
                    required: [label, value]
                    properties:
                      label: { type: string }
                      value: { type: string }
                nextCursor: { type: [string, 'null'] }